+++ stylesheets/_session_theme_dark.scss
@@ -21,6 +21,9 @@
     color: $session-color-dark-5;
   }
 
+  .module-message__container {
+    box-shadow: 0px 0px 8px 2px rgba(0, 0, 0, 0.69);
+  }
   .module-message__container--incoming {
     background-color: $session-color-dark-1;
   }
